➕ ADD: Proveedores app
The app "Proveedores" was added to the project. This contains a model
file and an API folder with serializers, view sets, and a router.

The settings file now contains three separate lists of apps names

Now the project contains a BASH script to update and launch the app on
Docker

Issue: SGI-AC-2